IEP-449: Code enhancement and review comments.
