+++ bundles/com.espressif.idf.ui/.classpath
@@ -3,10 +3,7 @@
 	<classpathentry kind="src" path="src"/>
 	<classpathentry exported="true" kind="lib" path="lib/commonmark-0.16.1.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/ini4j-0.5.4.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/commons-collections4-4.4.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/commons-io-2.9.0.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/commons-lang3-3.12.0.jar"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-11">
 		<attributes>
 			<attribute name="module" value="true"/>
 		</attributes>
