+++ bundles/com.espressif.idf.ui/META-INF/MANIFEST.MF
@@ -37,7 +37,4 @@ Export-Package: com.espressif.idf.ui,
  com.espressif.idf.ui.dialogs
 Bundle-ClassPath: .,
  lib/commonmark-0.16.1.jar,
- lib/ini4j-0.5.4.jar,
- lib/commons-collections4-4.4.jar,
- lib/commons-io-2.9.0.jar,
- lib/commons-lang3-3.12.0.jar
+ lib/ini4j-0.5.4.jar
