+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/menuitem/LanguageDynamicMenuItem.java
@@ -3,50 +3,76 @@
  * Use is subject to license terms.
  *******************************************************************************/
 
-
 package com.espressif.idf.ui.menuitem;
 
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.jface.action.ContributionItem;
 import org.eclipse.swt.SWT;
+import org.eclipse.swt.events.SelectionAdapter;
+import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.MenuItem;
+import org.eclipse.ui.PlatformUI;
 
-import com.espressif.idf.ui.menuitem.listener.LanguageItemSelectionListener;
+import com.espressif.idf.core.logging.Logger;
+import com.espressif.idf.core.util.EclipseIniUtil;
 
 public class LanguageDynamicMenuItem extends ContributionItem
 {
-	private LanguageItemSelectionListener languageItemSelectionListener;
+	private static final String LANGUAGE_SWITCH = "-nl"; //$NON-NLS-1$
+	private EclipseIniUtil eclipseIniUtil;
 
 	public LanguageDynamicMenuItem()
 	{
-		languageItemSelectionListener = new LanguageItemSelectionListener();
-	}
-
-	public LanguageDynamicMenuItem(String id)
-	{
-		super(id);
-		languageItemSelectionListener = new LanguageItemSelectionListener();
+		try
+		{
+			eclipseIniUtil = new EclipseIniUtil();
+		}
+		catch (Exception e)
+		{
+			Logger.log(e);
+		}
 	}
 
 	@Override
 	public void fill(Menu menu, int index)
 	{
 		IConfigurationElement[] configElements = Platform.getExtensionRegistry()
-				.getConfigurationElementsFor("com.espressif.idf.ui.extensionspoint.languagepoint");
+				.getConfigurationElementsFor("com.espressif.idf.ui.locale.support");
 		for (IConfigurationElement iConfigurationElement : configElements)
 		{
 			MenuItem menuItem = new MenuItem(menu, SWT.PUSH);
-			String langCode = iConfigurationElement.getAttribute("id");
-			String langName = iConfigurationElement.getAttribute("name");
-			menuItem.setText(langName + " - " + langCode);
-			if (languageItemSelectionListener == null)
+			String langCode = iConfigurationElement.getAttribute("id"); //$NON-NLS-1$
+			String langName = iConfigurationElement.getAttribute("name"); //$NON-NLS-1$
+			menuItem.setText(langName);
+
+			menuItem.addSelectionListener(new SelectionAdapter()
 			{
-				languageItemSelectionListener = new LanguageItemSelectionListener();
-			}
-			
-			menuItem.addSelectionListener(languageItemSelectionListener);
+				@Override
+				public void widgetSelected(SelectionEvent selectionEvent)
+				{
+					if (eclipseIniUtil != null)
+					{
+						try
+						{
+							if (eclipseIniUtil.containsEclipseSwitchInEclipseIni(LANGUAGE_SWITCH))
+							{
+								if (!eclipseIniUtil.getEclipseIniSwitchValue(LANGUAGE_SWITCH).equals(langCode))
+								{
+									eclipseIniUtil.setEclipseSwitchValue(LANGUAGE_SWITCH, langCode);
+									PlatformUI.getWorkbench().restart(false);
+								}
+							}
+						}
+						catch (Exception e)
+						{
+							Logger.log(e);
+						}
+
+					}
+				}
+			});
 		}
 	}
